+++ api/src/db/drizzleTables.ts
@@ -1,0 +1,5 @@
+/**
+ * Re-export generated Drizzle schema so API code imports from `./db/drizzleTables`
+ * instead of deep paths under `types/generated`.
+ */
+export * from "../../../types/generated/ts/db/schema.generated";
