+++ scripts/generate-types.ts
@@ -1,52 +1,24 @@
 import { compile } from "json-schema-to-typescript";
-import { mkdir, readdir, readFile, rm, writeFile } from "node:fs/promises";
+import { mkdir, readdir, rm, writeFile } from "node:fs/promises";
 import { join, relative } from "node:path";
-import { spawn } from "node:child_process";
 import {
 	OUT_SWIFT,
 	OUT_TS,
 	REPO_ROOT,
 	SCHEMA_DIR,
 	TYPES_ROOT,
+	appendLinesToGeneratedFile,
 	loadJson,
 	runMain,
 	schemaPathToSwiftTypeName,
 	schemaPathToTsName,
+	spawnExitOk,
 } from "./types-generation-utils.js";
 
 const COMMON_SCHEMA_ROOT_REF: Record<string, string> = {
 	"common/json": "#/$defs/JSONValue",
 	"common/rpc": "#/$defs/IdFilter",
 };
-
-async function appendLinesToGeneratedFile(
-	outPath: string,
-	lines: string[],
-): Promise<void> {
-	const current = await readFile(outPath, "utf-8");
-	await writeFile(
-		outPath,
-		`${current.trimEnd()}\n\n${lines.join("\n")}\n`,
-		"utf-8",
-	);
-}
-
-function spawnExitOk(
-	command: string,
-	args: string[],
-	options: { cwd: string; stdio?: "inherit" },
-	errorLabel: string,
-): Promise<void> {
-	return new Promise((resolve, reject) => {
-		const proc = spawn(command, args, options);
-		proc.on("exit", (code) =>
-			code === 0
-				? resolve()
-				: reject(new Error(`${errorLabel} exited ${code}`)),
-		);
-		proc.on("error", reject);
-	});
-}
 
 type LoadedSchemaFile = {
 	schemaPath: string;
